+++ NC/src/main/java/org/example/NC/repos/TariffСategoryRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TariffСategoryRepo extends CrudRepository<TariffСategory, Long> {
+public interface TariffСategoryRepo extends CrudRepository<TariffСategory, Integer> {
         List<TariffСategory> findByName(String name);
 }
