+++ NC/src/main/java/org/example/NC/repos/UserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<Client, Long> {
+public interface UserRepo extends JpaRepository<Client, Integer> {
    // Client findByLogin(String login);
 
     Client findByUsername(String username);
